refactor(icon): update all IgxIconModule references
Following the feature implementation of the IgxIconService and the way
the module is exported, the IgxIconModule references needs to be updated
in all components that depend on it.

BREAKING CHANGE: IgxIconModule should now be imported from lib/icon, instead of lib/icon/icon.component.
