+++ src/calendar/calendar.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
-import { IgxIconModule } from "../icon/icon.component";
+import { IgxIconModule } from "../icon";
 import { IgxCalendarComponent } from "./calendar.component";
 import {
     IgxCalendarDateDirective,
@@ -30,4 +30,4 @@ import {
     ],
     imports: [CommonModule, FormsModule, IgxIconModule]
 })
-export class IgxCalendarModule {}
+export class IgxCalendarModule { }
