+++ src/grid/index.ts
@@ -7,72 +7,72 @@ import { IgxForOfModule } from "../directives/for-of/for_of.directive";
 import { IgxInputModule } from "../directives/input/input.directive";
 import { IgxRippleModule } from "../directives/ripple/ripple.directive";
 import { IgxToggleModule } from "../directives/toggle/toggle.directive";
-import { IgxIconModule } from "../icon/icon.component";
+import { IgxIconModule } from "../icon";
 import { IgxGridAPIService } from "./api.service";
 import { IgxGridCellComponent } from "./cell.component";
 import { IgxColumnComponent } from "./column.component";
 import { IgxGridFilterComponent } from "./grid-filtering.component";
 import { IgxGridHeaderComponent } from "./grid-header.component";
 import {
-  IgxCellEditorTemplateDirective,
-  IgxCellFooterTemplateDirective,
-  IgxCellHeaderTemplateDirective,
-  IgxCellTemplateDirective
+    IgxCellEditorTemplateDirective,
+    IgxCellFooterTemplateDirective,
+    IgxCellHeaderTemplateDirective,
+    IgxCellTemplateDirective
 } from "./grid.common";
 import { IgxGridComponent } from "./grid.component";
 import { IgxGridFilterConditionPipe, IgxGridFilteringPipe, IgxGridPagingPipe, IgxGridSortingPipe } from "./grid.pipes";
 import { IgxGridRowComponent } from "./row.component";
 
 @NgModule({
-  declarations: [
-    IgxGridCellComponent,
-    IgxColumnComponent,
-    IgxGridComponent,
-    IgxGridRowComponent,
-    IgxGridHeaderComponent,
-    IgxCellFooterTemplateDirective,
-    IgxCellHeaderTemplateDirective,
-    IgxCellEditorTemplateDirective,
-    IgxCellTemplateDirective,
-    IgxGridFilterComponent,
-    IgxGridSortingPipe,
-    IgxGridPagingPipe,
-    IgxGridFilteringPipe,
-    IgxGridFilterConditionPipe
-  ],
-  entryComponents: [
-    IgxColumnComponent
-  ],
-  exports: [
-    IgxGridComponent,
-    IgxGridCellComponent,
-    IgxGridRowComponent,
-    IgxColumnComponent,
-    IgxGridHeaderComponent,
-    IgxGridFilterComponent,
-    IgxCellFooterTemplateDirective,
-    IgxCellHeaderTemplateDirective,
-    IgxCellEditorTemplateDirective,
-    IgxCellTemplateDirective
-  ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    IgxButtonModule,
-    IgxDatePickerModule,
-    IgxIconModule,
-    IgxRippleModule,
-    IgxInputModule,
-    IgxToggleModule,
-    IgxForOfModule
-  ],
-  providers: [IgxGridAPIService]
+    declarations: [
+        IgxGridCellComponent,
+        IgxColumnComponent,
+        IgxGridComponent,
+        IgxGridRowComponent,
+        IgxGridHeaderComponent,
+        IgxCellFooterTemplateDirective,
+        IgxCellHeaderTemplateDirective,
+        IgxCellEditorTemplateDirective,
+        IgxCellTemplateDirective,
+        IgxGridFilterComponent,
+        IgxGridSortingPipe,
+        IgxGridPagingPipe,
+        IgxGridFilteringPipe,
+        IgxGridFilterConditionPipe
+    ],
+    entryComponents: [
+        IgxColumnComponent
+    ],
+    exports: [
+        IgxGridComponent,
+        IgxGridCellComponent,
+        IgxGridRowComponent,
+        IgxColumnComponent,
+        IgxGridHeaderComponent,
+        IgxGridFilterComponent,
+        IgxCellFooterTemplateDirective,
+        IgxCellHeaderTemplateDirective,
+        IgxCellEditorTemplateDirective,
+        IgxCellTemplateDirective
+    ],
+    imports: [
+        CommonModule,
+        FormsModule,
+        IgxButtonModule,
+        IgxDatePickerModule,
+        IgxIconModule,
+        IgxRippleModule,
+        IgxInputModule,
+        IgxToggleModule,
+        IgxForOfModule
+    ],
+    providers: [IgxGridAPIService]
 })
 export class IgxGridModule {
-  public static forRoot() {
-    return {
-      ngModule: IgxGridModule,
-      providers: [IgxGridAPIService]
-    };
-  }
+    public static forRoot() {
+        return {
+            ngModule: IgxGridModule,
+            providers: [IgxGridAPIService]
+        };
+    }
 }
